feat: typecsript - services refactor
