+++ server/controllers/admin.ts
@@ -1,16 +1,16 @@
-import { NavigationService, ToBeFixed } from '../../types';
+import { IClientService, ICommonService, ToBeFixed } from '../../types';
 import { getPluginService, parseParams } from '../utils';
 import { errorHandler } from '../utils';
 import { IAdminController } from '../../types'
 
 const adminControllers: IAdminController = {
-	getService(name = "navigation"): NavigationService {
-    return getPluginService(name);
+  getService<T = IClientService>(name = "admin") {
+    return getPluginService<T>(name);
   },
-	async get() {
-		return await this.getService().get();
-	},
-	post(ctx) {
+  async get() {
+    return await this.getService().get();
+  },
+  post(ctx) {
     const { auditLog } = ctx;
     const { body = {} } = ctx.request;
     return this.getService().post(body, auditLog);
@@ -21,10 +21,11 @@ const adminControllers: IAdminController = {
     const { body = {} } = ctx.request;
     return this.getService().put(id, body, auditLog)
       .catch(errorHandler(ctx));
-	},
-	async config() {
+  },
+  async config() {
     return this.getService().config();
-	},
+  },
+  
 	async updateConfig(ctx) {
     try {
       await this.getService().updateConfig(ctx.request.body);
@@ -45,17 +46,17 @@ const adminControllers: IAdminController = {
 
   async settingsConfig() {
     return this.getService().config(true);
-	},
-	
-	async settingsRestart(ctx) {
+  },
+
+  async settingsRestart(ctx) {
     try {
       await this.getService().restart();
       return ctx.send({ status: 200 });
     } catch (e: ToBeFixed) {
       errorHandler(ctx)(e);
     }
-	},
-	async getById(ctx) {
+  },
+  async getById(ctx) {
     const { params } = ctx;
     const { id } = parseParams(params);
     return this.getService().getById(id);
@@ -63,7 +64,7 @@ const adminControllers: IAdminController = {
   async getContentTypeItems(ctx) {
     const { params } = ctx;
     const { model } = parseParams(params);
-    return this.getService().getContentTypeItems(model)
+    return this.getService<ICommonService>('common').getContentTypeItems(model)
   },
 };
