+++ server/services/index.ts
@@ -1,7 +1,13 @@
 'use strict';
 
-import navigation from './navigation';
+import commonService from './common';
+import adminService from './admin';
+import clientService from './client';
 
-export = {
-  navigation
+const navigationService = {
+  common: commonService,
+  admin: adminService,
+  client: clientService,
 };
+
+export default navigationService;
