+++ server/utils/index.ts
@@ -1,1 +1,2 @@
 export * from './functions';
+export * from './constant';
